electron-17: refactored some files
